backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
sqlalchemy==2.0.36
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
